Add Side.NONE and Side.create
